requirements.txt
# Minimum Django and REST framework version
Django>=1.6
djangorestframework>=3.1

# Test requirements
pytest-django==2.8.0
pytest==2.6.4
pytest-cov==1.6
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
